requirements.txt
python-dotenv
responder
requests

